Refactoring slack notifier
